Properly set prefix and ident for plugin manager and core plugin.
For galog a prefix is a string to be added to all log entries in case the application needs an extra context (i.e. subsystems or in the case of guest-agent the distinction between plugin manager and core plugin - given they usually to the same logging bucket).

In case of syslog/journal the entries will look contain:
<ident>: [prefix]: <msg>

We are temporarily moving the SetPrefix() call to windows only where it makes sense for now - since right now we have two systemd units it would be redundant to have for example "google_guest_agent_manager: GCEPluginManager: <msg>". Whenever we have it consolidated into a single systemd unit it will look like "google_guest_agent: PluginManager: <msg>" and "google_guest_agent: CorePlugin: <msg>".

PiperOrigin-RevId: 697004609