Make sure to check through all dhclient processes before returning whether a dhclient process for a given NIC and IP version exists.
Given example where we're looking for an IPv6 process for a specific interface, if its IPv4 process is listed before its IPv6 process, then `containsProtocolArg` is false, and the function returns false erroneously without checking any of the other processes.

This changes it so that each process is checked until either a match is found or we fail to find a match among any of them.

PiperOrigin-RevId: 789343175