Fix routes setup attempting to add already-existing routes.
When a route of `x.x.x.x/32` or `x::x/128` is set in MDS, when it is added as a route, the CIDR notation is automatically removed. This means when we parse the route table again to check for whether a route exists, it fails to find this route due to comparing the CIDR notation `x.x.x.x/32` with the existing route `x.x.x.x`. To resolve this, we instead check if the provided route (if in CIDR notation) only contains one IP, and if so, we treat it as just an IP address rather than a CIDR address.

PiperOrigin-RevId: 820405866