Wait for systemd-networkd to acquire DHCP leases before finishing network configuration.
There are two CREATE events for each lease, whereas there's only one `WRITE` event. But we can also look for a RENAME event, since the leases are written to a temporary file first before being renamed to the final file. Either way, by the time the leases are written, they are already obtained, so it shouldn't matter whether we look at `RENAME` or `WRITE`.

PiperOrigin-RevId: 822786393