Add a DNS probe module to the core plugin.
This module implements a scheduled job that periodically probes the GCE DNS server (169.254.169.254). The probe queries a domain constructed to reflect the status of the *previous* probe (e.g., `nxdomain`, `timeout`, `error`). The module is disabled via the disable-dns-probe attribute. The probe enables platform level monitoring of the GCE DNS server health.

PiperOrigin-RevId: 825720407