Default IPv4 address mask to `/32`.
On Windows, these route masks are very important, and the `net` Golang library sets different masks (`/8`, `/16`, `/24`) for different IPs depending on the value of the first byte (if the first byte is less than `128`, then the IP gets a default mask of `/8`. If it's less than `192`, then it gets a default mask of `/16`).

This messes up the routes on Windows as these are extremely large IP ranges, and it caused issues in the recent Guest Agent package regarding WSFC. Forcing a default mask of `/32` for IPv4 should fix this issue.

This also has no effect on Linux, as the Linux implementation does not depend on these default masks.

PiperOrigin-RevId: 826189014