Add a plugin cleanup watcher.
Sometimes the plugin manager fails to remove plugins, but still reports back success to ACS. Some of the steps don't block the removal, but given the lingering resources, the failures will eventually culminate in the instance running out of space entirely.

This adds a daily cleanup functionality to the guest agent, where it runs at least once a day, and is also triggered by a restart of the guest agent. This helps ensure that those lingering resources are cleaned up.

PiperOrigin-RevId: 838899541