Make the reg library create registry keys if they don't exist.
Currently, if the key is missing, the guest agent will not try to create it. This leads to issues later on since the guest agent uses these keys to cache work it's already done -- doing duplicate work later on can cause a lot of issues, especially with regards to password resets.

This should prevent that from happening, and also allow for a slight optimization in only creating the key if it will be used.

PiperOrigin-RevId: 845483603