Address race condition between plugin creating its socket and the plugin manager trying to connect to it.
In some cases, the plugin is slow in starting up (perhaps due to the OS scheduler), which means that when the plugin manager attempts to connect to the plugin via a UDS unix socket, if the socket doesn't exist, then it will create the socket. This causes issues for the plugin, as it will also try to create the socket, but if the socket already exists, it will fail and then enter a crash loop.

As such, we add a check before connecting to make sure that the socket has been created by the plugin before attempting to connect to it (only for UDS connections). We can do this by constantly checking for the socket's existence with a timeout, and if we don't find it after the timeout, then we know something went wrong with the plugin.

PiperOrigin-RevId: 845995507