Log GDU Universe only once.
If we're running in GDU, the watcher will keep renewing, which means so long as we check for the universe, the log message will be logged on each renewal. This should force it to log just once, instead of every time.

We also don't expect the key to change while the guest agent is running.

PiperOrigin-RevId: 846842836